Add hook for actions to state.
The state machine needs to be able to handle any number of actions when
online. It can now accept an object of actions that are attached to the
online state hander. Next step will be to add an action object that
registers new actions and marshalls them into this object structure.